First draft
Subclasses of JLWorkflowItem should define for each step:

- (JLWorkflowStepStatus)statusFor<N>Step;
Called by the framework to find out if the step is waiting on
dependencies (UnmetDependency), ready to run (CanRun) or is
completed (Completed). This method should be quick to return
and should be idempotent.

One of:
- (NSArray *)perform<N>Step:(JLWorkflowSyncToken *)token;
- (void)perform<N>StepAsynchronously:(JLWorkflowAsyncToken *)token;

The first variant is for steps that will run synchronously (i.e.
the method won't return until the step completes), while the second
variant is for asynchronous steps (the method returns before the
step is completed, the step will complete at a later time)

In the first case, the step execution method should follow the
following pattern:
- Perform step processing, not altering any instance state
  - At any point, the token may be checked for validity by
    using the .valid property. If the token becomes invalid,
    the method should return immediately.
  - At any point, progress may be updated (0.0-1.0) by setting
    the token's .progress property. There is no need to check
    if the token is valid before updating the progress.
- Once processing is complete, if successful, the code may update
  instance state provided the token is still valid. There's an issue
  here as the token may still become invalid while state is being
  updated, this will be addressed in a later revision.
- Finally, the method should return an array of errors (or nil/empty
  array if no errors occurred).

In the async case, the step execution method should follow the
following pattern:
- Launch the asynchronous operation (NSURLConnection, performSelectorAfterDelay,
  etc.) and return.
- When the asynchronous operation completes (callback called, etc.)
  instance state should be updated after verifying the token is
  still valid. Again, there's a potential race condition that will
  be addressed later.
- Finally, the token's [completeAsynchronousStepWithErrors:] method
  should be called.
Since in the async case, the step will complete in a different scope
than when the step was started, it is important to save a reference
to the AsyncToken (so that progress can be updated and completion
can be notified). Some async facilities provide ways to do this,
by means of a context parameter or equivalent. However, other facilities
such as NSURLConnection, do not have this option. In those cases,
the [token associateWithObject:obj] can be very useful. Once this
method is called, subsequent [JLWorkflowToken associatedTokenForObject:]
will return the token.

Planned enhancements:
- Stop requiring the subclass to return a three-state status
  (UnmetDependency, CanRun, Completed). Instead, possibly just
  return a boolean indicating if the task may be run or not.
  The completion state should be kept around by the framework.
- Fix race conditions during instance state updates, possibly
  doing this by calling a token method with a block that will
  update the instance. This block would be run in the main thread.
- In that case, only use the token to notify success/failure,
  deprecate the sync method of return the errors.
- Rename classes, maybe using WF prefix (WFToken instead of JLWorkflowToken)
- Split classes into files