Simplify mayRun<N>Step method
Replaced statusFor<N>Step with mayRun<N>Step.
This method need only return a boolean indicating if the step
may be run or not based on dependencies.

Completion state for each step is now kept by the framework,
and assumed to be true when the step completes successfully,
false otherwise.