Updated step execution API
perform<N>Step and perform<N>StepAsynchronously may now return a BOOL.
Steps return NO to indicate that the step is already complete and no
processing was necessary. The framework will mark the step as completed
and move on. When returning NO, it is important to ensure that notifyCompletion
is not called on the token - the framework will abort() if this happens.

Steps return YES to indicate that the step was completed (successfully
or not, based on whether notifyCompletion was called with errors or not)
In this cases, the step may optionally have called notifyCompletion to
provide a completionBlock. Any state updates on the WFObject subclass
must be performed in a completionBlock, so the only case where not
calling notifyCompletion makes sense is if no state needs to be updated.

If the step chooses not to return BOOL, it is assumed that the step
was completed.