Prevent serialization of long running sync steps
This would happen when a sync step was run while previous canceled
instances were still executing. A single NSInvocation was used
for all executions of a given sync step. Execution had to then
by synchronized on this NSInvocation, which then caused the serialization.

Now a different NSInvocation is created every time a sync step must be run.