Implement WFObject & WFStep NSCoding
WFObject implements NSCoding. But WFStep doesn't. Why?

First, why should it? It should because WFStep has data that must
be persisted. At least the array of errors should persist from
session to session. And possibly the completed flag, although
that one can be regenerated by attempting to run all the steps that
are ready.

Ok, so it should. But why doesn't it?

It doesn't because as WFObject subclasses change, we want the
WFObject steps member to track the actual subclass implementation.
So we can't just deserialize the steps array, as we would be
bringing potentially old metadata information (steps that don't
exist anymore while missing newly defined steps).

So the approach taken is to have WFObject create a dictionary
that will contain an entry per currently defined step. In each
entry, it will request the step to serialize whatever information
it deems to require persistence. On deserialization, WFObject
builds the steps array as usual (based on the current subclass
implementation) and for each available entry in the serialized
dictionary it asks the appropriate step (keyed by name) to
restore the persisted values. This way, new steps still get
created, but don't deserialize any data - they remain in the
default state. While steps that were removed from the implementation
simply don't claim the previously persisted values and are
silently ignored.

There may be a way to do this via regular NSCoding, but I'm
unaware of it.